+++ CHANGELOG.md
@@ -15,7 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   @janisslsm
 - Added loading payload from file
 - Added read8/read16/write8/write16 functions
-- Added 8.50 and 8.52 support
+- Added 8.50-9.60 support
+  - Initial 9.00-9.60 ROP chain, by @janisslsm
+- Added GitHub actions to build PRs, push to `main`, and tags for releases.
 
 ### Fixed